+++ src/Couchbase.EntityFrameworkCore/Query/Internal/CouchbaseParameterBasedSqlProcessorFactory.cs
@@ -1,0 +1,10 @@
+using Microsoft.EntityFrameworkCore.Query;
+
+namespace Couchbase.EntityFrameworkCore.Query.Internal;
+
+public class CouchbaseParameterBasedSqlProcessorFactory(RelationalParameterBasedSqlProcessorDependencies dependencies)
+    : IRelationalParameterBasedSqlProcessorFactory
+{
+    public virtual RelationalParameterBasedSqlProcessor Create(RelationalParameterBasedSqlProcessorParameters parameters)
+        => new CouchbaseParameterBasedSqlProcessor(dependencies, parameters);
+}